+++ rust-client/src/instructions/dynamic_amm/deposit.rs
@@ -51,7 +51,7 @@ pub fn process_deposit_dynamic_pool(args: &Args, sub_args: &DepositDynamicAmmArg
                 &keypair.pubkey(),
                 &keypair.pubkey(),
                 &pool_state.lp_mint,
-                &spl_token::ID,
+                &anchor_spl::token::ID,
             ),
         );
     }
@@ -90,7 +90,7 @@ pub fn process_deposit_dynamic_pool(args: &Args, sub_args: &DepositDynamicAmmArg
             user_b_token: get_associated_token_address(&keypair.pubkey(), &pool_state.token_b_mint),
             user: keypair.pubkey(),
             vault_program: prog_dynamic_vault::ID,
-            token_program: spl_token::ID,
+            token_program: anchor_spl::token::ID,
         }
         .to_account_metas(None),
         data: prog_dynamic_amm::instruction::AddBalanceLiquidity {