+++ ts-client/src/amm/tests/initializeCustomizablePermissionlessConstantProductPool.test.ts
@@ -4,8 +4,13 @@ import { airDropSol, getOrCreateATA, mockWallet, wrapSol } from './utils';
 import { createMint, mintTo, NATIVE_MINT } from '@solana/spl-token';
 import AmmImpl from '..';
 import { BN } from 'bn.js';
-import { ActivationType } from '../types';
-import { createProgram, deriveCustomizablePermissionlessConstantProductPoolAddress } from '../utils';
+import { ActivationType, FeeCurvePoints } from '../types';
+import {
+  createProgram,
+  deriveCustomizablePermissionlessConstantProductPoolAddress,
+  FeeCurveInfo,
+  FeeCurveType,
+} from '../utils';
 
 const connection = new Connection('http://127.0.0.1:8899', 'confirmed');
 const provider = new AnchorProvider(connection, mockWallet, {
@@ -50,6 +55,33 @@ describe('Initialize customizable permissionless constant product pool', () => {
     const tokenAAmount = new BN(1_000_000);
     const tokenBAmount = new BN(1_000_000);
 
+    const feePoints: FeeCurvePoints = [
+      {
+        feeBps: 1300,
+        activatedDuration: 5,
+      },
+      {
+        feeBps: 1100,
+        activatedDuration: 10,
+      },
+      {
+        feeBps: 900,
+        activatedDuration: 15,
+      },
+      {
+        feeBps: 700,
+        activatedDuration: 20,
+      },
+      {
+        feeBps: 500,
+        activatedDuration: 25,
+      },
+      {
+        feeBps: 400,
+        activatedDuration: 30,
+      },
+    ];
+
     const initializeTx = await AmmImpl.createCustomizablePermissionlessConstantProductPool(
       connection,
       mockWallet.publicKey,
@@ -59,11 +91,11 @@ describe('Initialize customizable permissionless constant product pool', () => {
       tokenBAmount,
       {
         // Denominator is default to 100_000
-        tradeFeeNumerator: 2500,
+        tradeFeeNumerator: 15_000,
         activationType: ActivationType.Timestamp,
         activationPoint: null,
         hasAlphaVault: false,
-        padding: Array(90).fill(0),
+        feeCurve: FeeCurveInfo.flat(feePoints),
       },
     );
 
@@ -79,6 +111,26 @@ describe('Initialize customizable permissionless constant product pool', () => {
     );
 
     const pool = await AmmImpl.create(connection, poolKey);
-    expect(pool.feeBps.eq(new BN(250))).toBeTruthy();
+    const feeBps = pool.poolState.fees.tradeFeeNumerator.muln(10000).div(pool.poolState.fees.tradeFeeDenominator);
+    expect(feeBps.eq(new BN(1500))).toBeTruthy();
+    expect(JSON.stringify(pool.poolState.feeCurve.feeCurveType)).toBe(JSON.stringify(FeeCurveType.flat()));
+
+    for (let i = 0; i < feePoints.length; i++) {
+      // 0 is initialize fee rate
+      const p1 = pool.poolState.feeCurve.points[i + 1];
+      const p2 = feePoints[i];
+
+      expect(p1.feeBps).toBe(p2.feeBps);
+      expect(p1.activatedPoint.toNumber()).toBe(
+        pool.poolState.bootstrapping.activationPoint.toNumber() + p2.activatedDuration,
+      );
+    }
+
+    const beforeFee = pool.feeBps;
+    await new Promise((res) => setTimeout(res, 7000));
+    await pool.updateState();
+    const afterFee = pool.feeBps;
+
+    expect(afterFee.toNumber()).toBeLessThan(beforeFee.toNumber());
   });
 });