+++ CHANGELOG.md
@@ -19,6 +19,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## dynamic-amm-quote [0.0.5] - PR[#204](https://github.com/MeteoraAg/dynamic-amm-sdk/pull/204/)
+
+### Changed
+
+- Support fee curve quote
+
+## dynamic-amm [0.6.2] - PR[#204](https://github.com/MeteoraAg/dynamic-amm-sdk/pull/204/)
+
+- Add fee curve related state in `Pool` account
+
+## @meteora-ag/dynamic-amm-sdk [1.3.2] - PR[#204](https://github.com/MeteoraAg/dynamic-amm-sdk/pull/204/)
+
+### Changed
+
+- improve stake pool price
+- `createCustomizablePermissionlessConstantProductPool` function `customizableParams` now require `feeCurve`. See [test](./ts-client/src/amm/tests/initializeCustomizablePermissionlessConstantProductPool.test.ts) example.
+- `createCustomizablePermissionlessConstantProductPool` function `customizableParams` no longer need to pass in `padding`
+- `calculateTradingFee` and `calculateProtocolTradingFee` function receive `PoolFees` instead of `PoolState`
+
 ## commons [0.0.5] - PR[#195](https://github.com/MeteoraAG/dynamic-amm-sdk/pull/195)
 
 ### Added